roachtest: break apart mixed_version_backup
This commit breaks apart the `mixed_version_backup.go` file into three
separate files.

Epic: None

Release note: None
