+++ pkg/cmd/roachtest/tests/BUILD.bazel
@@ -36,7 +36,9 @@ go_library(
         "awsdms.go",
         "backup.go",
         "backup_fixtures.go",
+        "backup_restore_driver.go",
         "backup_restore_roundtrip.go",
+        "backup_restore_utils.go",
         "buffered_logging.go",
         "build_info.go",
         "canary.go",
